Add support for appending Brekel armature data block.
Previously, one was required to append a Brekel armature object, not
simply the armature datablock, in order to support correct re-targeting
of Brekel animation actions. This is because appending the armature
datablock set the bone rotation mode to quaternion, not Euler by
default. But Brekel actions assume Euler rotation.

This modification resets bone rotation to Euler meaning that even if
armature bones are initially set to quaternion (which is the default for
an imported armature datablock), they will be reset to Euler. This is a
no-op for bones that are already set to Euler.